[Food][Fix] 버그 수정
- 가격 null 허용

+++ src/main/java/com/bobeat/backend/domain/store/external/kakao/dto/KakaoMenuDto.java
@@ -3,5 +3,5 @@ package com.bobeat.backend.domain.store.external.kakao.dto;
 import lombok.Builder;
 
 @Builder
-public record KakaoMenuDto(String imageUrl, String name, int price) {
+public record KakaoMenuDto(String imageUrl, String name, Long price) {
 }
